Update README.md
The YouTube Clone App is a Flutter-based application that replicates some of the core functionalities of the popular video-sharing platform, YouTube. It allows users to search for and view YouTube videos within the app. The app uses the youtube_explode_dart package to interact with the YouTube API for video search and retrieval.